Remove duplicated colander tests
